Move `setNativeProps` to `platformFunctions`
