+++ src/reanimated2/platformFunctions/index.ts
@@ -1,4 +1,6 @@
+'use strict';
 export { dispatchCommand } from './dispatchCommand';
 export { measure } from './measure';
 export { scrollTo } from './scrollTo';
 export { setGestureState } from './setGestureState';
+export { setNativeProps } from './setNativeProps';
